refactor: Reorganise Chapter1_Escape dir
