+++ .gitignore
@@ -1,0 +1,2 @@
+**/data_dumps
+**/__pycache__/
