+++ Chapter1_Escape/bag.py
@@ -1,7 +1,8 @@
 # Uses a turtle instance to draw a bag.
 
+
 def draw_bag(t):
-    t.pen(pencolor='brown', pensize=5)
+    t.pen(pencolor="brown", pensize=5)
     t.penup()
     t.goto(-35, 35)
     t.pendown()
