+++ Chapter1_Escape/draw_line.py
@@ -3,6 +3,7 @@ from .escape import has_escaped
 # Uses a turtle to draw a line until
 # the turtles position is outside of the bag
 
+
 def draw_line(t):
     angle = 0
     step = 5
