+++ Chapter1_Escape/escape.py
@@ -1,6 +1,7 @@
-# Use the turtles position 
-# to check whether it is outside 
+# Use the turtles position
+# to check whether it is outside
 # of the bag
+
 
 def has_escaped(position):
     x = int(position[0])
