+++ Chapter1_Escape/store_position_data.py
@@ -3,6 +3,7 @@ from .escape import has_escaped
 # Store the previous positions of the turtle
 # in the provided history list
 
+
 def store_position_data(t, history):
     [posX, posY] = t.position()
     history.append([posX, posY, has_escaped([posX, posY])])
